+++ Memoria.Launcher/Launcher/SettingsGrid_Cheats.cs
@@ -19,7 +19,7 @@ using HorizontalAlignment = System.Windows.HorizontalAlignment;
 
 namespace Memoria.Launcher
 {
-    public sealed class SettingsGrid_Cheats : UiGrid, INotifyPropertyChanged
+    public sealed class SettingsGrid_Cheats : Settings
     {
         public SettingsGrid_Cheats()
         {
@@ -46,9 +46,6 @@ namespace Memoria.Launcher
 
             row++;
             row++;
-            /*
-            AddUiElement(UiTextBlockFactory.Create("──────────────────────────────────────"), row, 0, 1, 8).Foreground = Brushes.White;
-            */
 
             CheckBox stealingAlwaysWorks = AddUiElement(UiCheckBoxFactory.Create(Lang.Settings.MaxStealRate, null), row, 0, 1, 8);
             stealingAlwaysWorks.SetBinding(ToggleButton.IsCheckedProperty, new Binding(nameof(StealingAlwaysWorks)) { Mode = BindingMode.TwoWay });
@@ -201,471 +198,5 @@ namespace Memoria.Launcher
             LoadSettings();
         }
 
-        public Int16 StealingAlwaysWorks
-        {
-            get { return _stealingalwaysworks; }
-            set
-            {
-                if (_stealingalwaysworks != value)
-                {
-                    _stealingalwaysworks = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
-        public Int16 NoAutoTrance
-        {
-            get { return _noautotrance; }
-            set
-            {
-                if (_noautotrance != value)
-                {
-                    _noautotrance = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
-        public Int16 GarnetConcentrate
-        {
-            get { return _garnetconcentrate; }
-            set
-            {
-                if (_garnetconcentrate != value)
-                {
-                    _garnetconcentrate = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
-        public Int16 BreakDamageLimit
-        {
-            get { return _breakDamageLimit; }
-            set
-            {
-                if (_breakDamageLimit != value)
-                {
-                    _breakDamageLimit = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
-        public Int16 AccessBattleMenu
-        {
-            get { return _accessBattleMenu; }
-            set
-            {
-                if (_accessBattleMenu != value)
-                {
-                    _accessBattleMenu = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
-        public String AvailableBattleMenus => AccessBattleMenu < 3 ? " \"Equip\", \"SupportingAbility\"" : "";
-        public Int16 SpeedMode
-        {
-            get { return _speedmode; }
-            set
-            {
-                if (_speedmode != value)
-                {
-                    _speedmode = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
-        public Int16 SpeedFactor
-        {
-            get { return _speedfactor; }
-            set
-            {
-                if (_speedfactor != value)
-                {
-                    _speedfactor = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
-        public Int16 BattleTPS
-        {
-            get { return _battletpsfactor; }
-            set
-            {
-                if (_battletpsfactor != value)
-                {
-                    _battletpsfactor = value;
-                    BattleTPSDividedBy10 = ((Decimal)value / 15);
-                    OnPropertyChanged();
-                }
-            }
-        }
-        public Decimal BattleTPSDividedBy10
-        {
-            get { return ((Decimal)BattleTPS / 15); }
-            set
-            {
-                OnPropertyChanged();
-            }
-        }
-        public Int16 BattleAssistance
-        {
-            get { return _battleassistance; }
-            set
-            {
-                if (_battleassistance != value)
-                {
-                    _battleassistance = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
-        public Int16 Attack9999
-        {
-            get { return _attack9999; }
-            set
-            {
-                if (_attack9999 != value)
-                {
-                    _attack9999 = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
-        public Int16 NoRandomEncounter
-        {
-            get { return _norandomencounter; }
-            set
-            {
-                if (_norandomencounter != value)
-                {
-                    _norandomencounter = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
-        public Int16 MasterSkill
-        {
-            get { return _masterskill; }
-            set
-            {
-                if (_masterskill != value)
-                {
-                    _masterskill = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
-        public Int16 MaxCardCount
-        {
-            get { return _maxcardcount; }
-            set
-            {
-                if (_maxcardcount != value)
-                {
-                    _maxcardcount = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
-        public Int16 ReduceRandom
-        {
-            get { return _reducerandom; }
-            set
-            {
-                if (_reducerandom != value)
-                {
-                    _reducerandom = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
-
-        private Int16 _stealingalwaysworks, _noautotrance, _garnetconcentrate, _breakDamageLimit, _accessBattleMenu, _speedmode, _speedfactor, _battletpsfactor, _battleassistance, _attack9999, _norandomencounter, _masterskill, _maxcardcount, _reducerandom;
-
-        private readonly String _iniPath = AppDomain.CurrentDomain.BaseDirectory + "\\Memoria.ini";
-
-        private void LoadSettings()
-        {
-
-            try
-            {
-                IniFile iniFile = new IniFile(_iniPath);
-
-                String value = iniFile.ReadValue("Hacks", nameof(StealingAlwaysWorks));
-                if (String.IsNullOrEmpty(value))
-                {
-                    value = " 0";
-                    OnPropertyChanged(nameof(StealingAlwaysWorks));
-                }
-                if (!Int16.TryParse(value, out _stealingalwaysworks))
-                    _stealingalwaysworks = 0;
-
-
-                value = iniFile.ReadValue("Battle", nameof(NoAutoTrance));
-                if (String.IsNullOrEmpty(value))
-                {
-                    value = " 0";
-                    OnPropertyChanged(nameof(NoAutoTrance));
-                }
-                if (!Int16.TryParse(value, out _noautotrance))
-                    _noautotrance = 0;
-
-                value = iniFile.ReadValue("Battle", nameof(GarnetConcentrate));
-                if (String.IsNullOrEmpty(value))
-                {
-                    value = " 0";
-                    OnPropertyChanged(nameof(GarnetConcentrate));
-                }
-                if (!Int16.TryParse(value, out _garnetconcentrate))
-                    _garnetconcentrate = 0;
-
-                value = iniFile.ReadValue("Battle", nameof(BreakDamageLimit));
-                if (String.IsNullOrEmpty(value))
-                {
-                    value = " 0";
-                    OnPropertyChanged(nameof(BreakDamageLimit));
-                }
-                if (!Int16.TryParse(value, out _breakDamageLimit))
-                    _breakDamageLimit = 0;
-
-                value = iniFile.ReadValue("Battle", "AccessMenus");
-                if (String.IsNullOrEmpty(value))
-                {
-                    value = " 0";
-                    OnPropertyChanged(nameof(AccessBattleMenu));
-                }
-                if (!Int16.TryParse(value, out _accessBattleMenu))
-                    _accessBattleMenu = 0;
-
-                value = iniFile.ReadValue("Cheats", nameof(SpeedMode));
-                if (String.IsNullOrEmpty(value))
-                {
-                    value = " 0";
-                    OnPropertyChanged(nameof(SpeedMode));
-                }
-                if (!Int16.TryParse(value, out _speedmode))
-                    _speedmode = 0;
-
-                value = iniFile.ReadValue("Cheats", nameof(SpeedFactor));
-                if (String.IsNullOrEmpty(value))
-                {
-                    value = " 2";
-                    OnPropertyChanged(nameof(SpeedFactor));
-                }
-                if (!Int16.TryParse(value, out _speedfactor))
-                    _speedfactor = 2;
-
-                value = iniFile.ReadValue("Graphics", " " + nameof(BattleTPS));
-                if (String.IsNullOrEmpty(value))
-                {
-                    _battletpsfactor = 15;
-                    value = " 15";
-                    OnPropertyChanged(nameof(BattleTPS));
-                }
-                if (!Int16.TryParse(value, out _battletpsfactor))
-                    _battletpsfactor = 15;
-                Boolean valueexists = decimal.TryParse(value, out Decimal decvalue);
-                BattleTPSDividedBy10 = valueexists ? decvalue / 15 : (decimal)1.5;
-
-                value = iniFile.ReadValue("Cheats", nameof(BattleAssistance));
-                if (String.IsNullOrEmpty(value))
-                {
-                    value = " 0";
-                    OnPropertyChanged(nameof(BattleAssistance));
-                }
-                if (!Int16.TryParse(value, out _battleassistance))
-                    _battleassistance = 0;
-
-                value = iniFile.ReadValue("Cheats", nameof(Attack9999));
-                if (String.IsNullOrEmpty(value))
-                {
-                    value = " 0";
-                    OnPropertyChanged(nameof(Attack9999));
-                }
-                if (!Int16.TryParse(value, out _attack9999))
-                    _attack9999 = 0;
-
-                value = iniFile.ReadValue("Cheats", nameof(NoRandomEncounter));
-                if (String.IsNullOrEmpty(value))
-                {
-                    value = " 0";
-                    OnPropertyChanged(nameof(NoRandomEncounter));
-                }
-                if (!Int16.TryParse(value, out _norandomencounter))
-                    _norandomencounter = 0;
-
-                value = iniFile.ReadValue("Cheats", nameof(MasterSkill));
-                if (String.IsNullOrEmpty(value))
-                {
-                    value = " 0";
-                    OnPropertyChanged(nameof(MasterSkill));
-                }
-                if (!Int16.TryParse(value, out _masterskill))
-                    _masterskill = 0;
-
-                /*value = null;
-                foreach (String prop in new String[] { "BattleFPS", "FieldFPS", "WorldFPS" })
-                {
-                    value = iniFile.ReadValue("Graphics", prop);
-                    if (!String.IsNullOrEmpty(value))
-                        break;
-                }*/
-
-                value = iniFile.ReadValue("TetraMaster", nameof(MaxCardCount));
-                if (String.IsNullOrEmpty(value))
-                {
-                    value = " 100";
-                    OnPropertyChanged(nameof(MaxCardCount));
-                }
-                Int16 cardCount = 100;
-                if (!Int16.TryParse(value, out cardCount))
-                    _maxcardcount = 0;
-                else
-                    _maxcardcount = cardCount < 9999 ? (Int16)0 : (Int16)1;
-
-
-                value = iniFile.ReadValue("TetraMaster", nameof(ReduceRandom));
-                if (String.IsNullOrEmpty(value))
-                {
-                    value = " 0";
-                    OnPropertyChanged(nameof(ReduceRandom));
-                }
-                if (!Int16.TryParse(value, out _reducerandom))
-                    _reducerandom = 0;
-
-                Refresh(nameof(StealingAlwaysWorks));
-                Refresh(nameof(NoAutoTrance));
-                Refresh(nameof(GarnetConcentrate));
-                Refresh(nameof(BreakDamageLimit));
-                Refresh(nameof(AccessBattleMenu));
-                Refresh(nameof(SpeedMode));
-                Refresh(nameof(SpeedFactor));
-                Refresh(nameof(BattleTPS));
-                Refresh(nameof(BattleAssistance));
-                Refresh(nameof(Attack9999));
-                Refresh(nameof(NoRandomEncounter));
-                Refresh(nameof(MasterSkill));
-                Refresh(nameof(MaxCardCount));
-                Refresh(nameof(ReduceRandom));
-            }
-            catch (Exception ex) { UiHelper.ShowError(Application.Current.MainWindow, ex); }
-        }
-
-        public event PropertyChangedEventHandler PropertyChanged;
-
-        private async void Refresh([CallerMemberName] String propertyName = null)
-        {
-            try
-            {
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-            }
-            catch (Exception ex)
-            {
-                UiHelper.ShowError(Application.Current.MainWindow, ex);
-            }
-        }
-
-        private async void OnPropertyChanged([CallerMemberName] String propertyName = null)
-        {
-            try
-            {
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-
-                IniFile iniFile = new IniFile(_iniPath);
-                switch (propertyName)
-                {
-                    case nameof(StealingAlwaysWorks):
-                        iniFile.WriteValue("Hacks", propertyName + " ", " " + StealingAlwaysWorks);
-                        if (StealingAlwaysWorks == 0)
-                        {
-                            iniFile.WriteValue("Hacks", propertyName + " ", " 0");
-                        }
-                        else if (StealingAlwaysWorks == 1)
-                        {
-                            iniFile.WriteValue("Hacks", "Enabled ", " 1");
-                            iniFile.WriteValue("Hacks", propertyName + " ", " 2");
-                        }
-                        break;
-                    case nameof(NoAutoTrance):
-                        iniFile.WriteValue("Battle", propertyName + " ", " " + NoAutoTrance);
-                        if (NoAutoTrance == 1)
-                            iniFile.WriteValue("Battle", "Enabled ", " 1");
-                        break;
-                    case nameof(GarnetConcentrate):
-                        iniFile.WriteValue("Battle", propertyName + " ", " " + GarnetConcentrate);
-                        if (GarnetConcentrate == 1)
-                            iniFile.WriteValue("Battle", "Enabled ", " 1");
-                        break;
-                    case nameof(BreakDamageLimit):
-                        iniFile.WriteValue("Battle", propertyName + " ", " " + BreakDamageLimit);
-                        if (BreakDamageLimit == 1)
-                            iniFile.WriteValue("Battle", "Enabled ", " 1");
-                        break;
-                    case nameof(AccessBattleMenu):
-                        iniFile.WriteValue("Battle", "AccessMenus ", " " + AccessBattleMenu);
-                        iniFile.WriteValue("Battle", "AvailableMenus ", AvailableBattleMenus);
-                        if (AccessBattleMenu > 0)
-                            iniFile.WriteValue("Battle", "Enabled ", " 1");
-                        break;
-                    case nameof(SpeedMode):
-                        iniFile.WriteValue("Cheats", propertyName + " ", " " + SpeedMode);
-                        if (SpeedMode == 1)
-                            iniFile.WriteValue("Cheats", "Enabled ", " 1");
-                        break;
-                    case nameof(SpeedFactor):
-                        if (SpeedFactor < 13)
-                            iniFile.WriteValue("Cheats", propertyName + " ", " " + SpeedFactor);
-                        break;
-                    case nameof(BattleTPS):
-                        iniFile.WriteValue("Graphics", propertyName + " ", " " + BattleTPS);
-                        if (BattleTPS != 15)
-                            iniFile.WriteValue("Cheats", "Enabled ", " 1");
-                        break;
-                    case nameof(BattleAssistance):
-                        iniFile.WriteValue("Cheats", propertyName + " ", " " + BattleAssistance);
-                        iniFile.WriteValue("Cheats", "Attack9999 ", " " + BattleAssistance); // Merged
-                        if (BattleAssistance == 1)
-                            iniFile.WriteValue("Cheats", "Enabled ", " 1");
-                        break;
-                    case nameof(Attack9999):
-                        iniFile.WriteValue("Cheats", propertyName + " ", " " + Attack9999);
-                        if (Attack9999 == 1)
-                            iniFile.WriteValue("Cheats", "Enabled ", " 1");
-                        break;
-                    case nameof(NoRandomEncounter):
-                        iniFile.WriteValue("Cheats", propertyName + " ", " " + NoRandomEncounter);
-                        if (NoRandomEncounter == 1)
-                            iniFile.WriteValue("Cheats", "Enabled ", " 1");
-                        break;
-                    case nameof(MasterSkill):
-                        iniFile.WriteValue("Cheats", propertyName + " ", " " + MasterSkill);
-                        iniFile.WriteValue("Cheats", "LvMax ", " " + MasterSkill);
-                        iniFile.WriteValue("Cheats", "GilMax ", " " + MasterSkill);
-                        if (MasterSkill == 1)
-                            iniFile.WriteValue("Cheats", "Enabled ", " 1");
-                        break;
-                    case nameof(MaxCardCount):
-                        if (MaxCardCount == 1)
-                        {
-                            iniFile.WriteValue("TetraMaster", propertyName + " ", " 10000");
-                            iniFile.WriteValue("TetraMaster", "Enabled ", " 1");
-                        }
-                        else if (MaxCardCount == 0)
-                        {
-                            iniFile.WriteValue("TetraMaster", propertyName + " ", " 100");
-                        }
-                        break;
-                    case nameof(ReduceRandom):
-                        if (ReduceRandom != 0)
-                            iniFile.WriteValue("TetraMaster", "Enabled ", " 1");
-                        iniFile.WriteValue("TetraMaster", propertyName + " ", " " + ReduceRandom);
-                        break;
-                }
-            }
-            catch (Exception ex)
-            {
-                UiHelper.ShowError(Application.Current.MainWindow, ex);
-            }
-        }
     }
 }
